NAV-28798: Flytter grafana initialisering til index.tsx
