+++ src/frontend/App.tsx
@@ -1,22 +1,19 @@
-import { useEffect } from 'react';
-
 import '@navikt/ds-css';
 import './index.css';
 
+import { AppProvider } from '@context/AppContext';
+import { AuthContextProvider } from '@context/AuthContext';
+import { HttpContextProvider } from '@context/HttpContext';
+import { ModalProvider } from '@context/ModalContext';
+import { SaksbehandlerProvider } from '@context/SaksbehandlerContext';
+import { FeatureTogglesProvider } from '@context/TogglesContext';
+import { useStartUmami } from '@hooks/useStartUmami';
+import { ErrorBoundary, ErrorBoundaryMedSaksbehandler } from '@komponenter/ErrorBoundary/ErrorBoundary';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
+import { erProd } from '@utils/miljø';
 
 import { Container } from './Container';
-import { AppProvider } from './context/AppContext';
-import { AuthContextProvider } from './context/AuthContext';
-import { HttpContextProvider } from './context/HttpContext';
-import { ModalProvider } from './context/ModalContext';
-import { SaksbehandlerProvider } from './context/SaksbehandlerContext';
-import { FeatureTogglesProvider } from './context/TogglesContext';
-import { useStartUmami } from './hooks/useStartUmami';
-import { ErrorBoundary, ErrorBoundaryMedSaksbehandler } from './komponenter/ErrorBoundary/ErrorBoundary';
-import { initGrafanaFaro } from './utils/grafanaFaro';
-import { erProd } from './utils/miljø';
 
 const queryClient = new QueryClient({
     defaultOptions: {
@@ -26,12 +23,8 @@ const queryClient = new QueryClient({
     },
 });
 
-const App = () => {
+export function App() {
     useStartUmami();
-
-    useEffect(() => {
-        initGrafanaFaro();
-    }, []);
 
     return (
         <ErrorBoundary>
@@ -55,6 +48,4 @@ const App = () => {
             </QueryClientProvider>
         </ErrorBoundary>
     );
-};
-
-export default App;
+}
