+++ src/frontend/index.tsx
@@ -1,26 +1,21 @@
 import { StrictMode } from 'react';
 
-import * as Sentry from '@sentry/browser';
 import { setDefaultOptions } from 'date-fns';
 import { nb } from 'date-fns/locale';
 import ReactDOM from 'react-dom';
 import { createRoot } from 'react-dom/client';
 
-import App from './App';
+import { App } from './App';
+import { initGrafanaFaro } from './grafana';
+import { initSentry } from './sentry';
 import { erLokal } from './utils/miljø';
 
 // Setter default locale til norsk bokmål for date-fns
 setDefaultOptions({ locale: nb });
 
-const environment = window.location.hostname;
-
 if (!erLokal()) {
-    Sentry.init({
-        dsn: 'https://10239ce4baed4db79d080d85f08b5878@sentry.gc.nav.no/26',
-        environment,
-        integrations: [Sentry.browserTracingIntegration()],
-        tracesSampleRate: 0.2,
-    });
+    initSentry();
+    initGrafanaFaro();
 }
 
 if (erLokal()) {
